refactor: move config ports and adapter to internal directory
