+++ internal/service-unit/application/core/initialization/configloader.go
@@ -2,13 +2,13 @@ package initialization
 
 import (
 	model "github.com/hanapedia/hexagon/pkg/api/v1"
-	"github.com/hanapedia/hexagon/pkg/operator/loader"
+	"github.com/hanapedia/hexagon/internal/config/application/ports"
 	l "github.com/hanapedia/hexagon/pkg/operator/logger"
 	"github.com/hanapedia/hexagon/pkg/operator/validation"
 )
 
 // GetConfig parses service-unit config via given formant
-func GetConfig(configLoader loader.ConfigLoader) model.ServiceUnitConfig {
+func GetConfig(configLoader ports.ConfigLoader) model.ServiceUnitConfig {
 	config, err := configLoader.Load()
 	if err != nil {
 		l.Logger.Fatalf("Error loading config: %v", err)
